+++ test/integration/sys_ssh/serverspec_chefdk30/localhost/ssh_spec.rb
@@ -1,7 +1,7 @@
 # Cookbook Name:: sys
 # Integration tests for recipe sys::ssh
 #
-# Copyright 2020-2022 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH
+# Copyright 2020-2024 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH
 #
 # Authors:
 #  Christopher Huhn   <c.huhn@gsi.de>
@@ -44,6 +44,10 @@ describe file('/etc/ssh/sshd_config') do
   its(:content) { should_not match(/^X11Forwarding yes$/) }
   # custom setting:
   its(:content) { should match(/^ClientAliveInterval 4711/) }
+
+  context 'bullseye or newer', if: debian_version >= 11 do
+    its(:content) { should match %r{^Include /etc/ssh/sshd_config.d/\*\.conf$} }
+  end
 end
 
 ### node['sys']['ssh']['ssh_config']: